Fixed a bug when posting a comment and refreshing the contact page the comment duplicates, and fixed delete note not working

+++ NCEA Website Internal/web development/Flask_Web/website/views.py
@@ -44,24 +44,21 @@ def contact():
             db.session.add(new_note)
             db.session.commit()
             flash('Comment Added!', category='success')
-            
-        
-        
+        return redirect(url_for('views.contact'))  # <-- This prevents duplicates on refresh
     return render_template("contact.html", user=current_user)
 
 
 
 #delete note route
 @views.route("/delete-note", methods=['POST'])
+@login_required
 def delete_note():
     note = json.loads(request.data)
     noteId = note['noteId']
     note = Note.query.get(noteId)
-    if note:
-        if note.user_id == current_user.id:
-            db.session.delete(note)
-            db.session.commit()
-            
+    if note and note.user_id == current_user.id:
+        db.session.delete(note)
+        db.session.commit()
     return jsonify({})
 
 @views.route('/scout')
